requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
psycopg2==2.9.9

# aiohttp==3.8.6
# aiosignal==1.3.1
# alembic==1.12.0
# async-timeout==4.0.3
# attrs==23.1.0
# blinker==1.6.2
# cachelib==0.10.2
# cachetools==5.3.1
# certifi==2023.7.22
# charset-normalizer==3.3.0
# click==8.1.7
# dnspython==2.4.2
# email-validator==2.0.0.post2
# facebook-sdk==3.1.0
# Flask==2.3.3
# Flask-Login==0.6.2
# Flask-Migrate==4.0.5
# Flask-Session==0.5.0
# Flask-SQLAlchemy==3.1.1
# Flask-WTF==1.2.1
# frozenlist==1.4.0
# google-ads==22.0.0
# google-api-core==2.12.0
# google-api-python-client==2.102.0
# google-auth==2.23.2
# google-auth-httplib2==0.1.1
# google-auth-oauthlib==1.1.0
# googleapis-common-protos==1.60.0
# greenlet==3.0.0
# grpcio==1.59.0
# grpcio-status==1.59.0
# httplib2==0.22.0
# idna==3.4
# itsdangerous==2.1.2
# Jinja2==3.1.2
# Mako==1.2.4
# MarkupSafe==2.1.3
# multidict==6.0.4
# numpy==1.26.0
# oauthlib==3.2.2
# openai==0.28.1
# pandas==2.1.1
# proto-plus==1.22.3
# protobuf==4.24.4
# psycopg2==2.9.9
# pyasn1==0.5.0
# pyasn1-modules==0.3.0
# pyparsing==3.1.1
# python-dateutil==2.8.2
# python-dotenv==1.0.0
# pytz==2023.3.post1
# PyYAML==6.0.1
# requests==2.31.0
# requests-oauthlib==1.3.1
# rsa==4.9
# six==1.16.0
# SQLAlchemy==2.0.21
# tqdm==4.66.1
# typing_extensions==4.8.0
# tzdata==2023.3
# uritemplate==4.1.1
# urllib3==2.0.6
# Werkzeug==2.3.7
# WTForms==3.0.1

